+++ raiden-ts/CHANGELOG.md
@@ -5,6 +5,7 @@
 - [#211] 'suggestPartners' method to fetch suggested partners from PFS
 - [#2417] Make 'start' async, introduce 'synced' promise, both resolves when syncing finishes
 - [#2444] Add adaptative sync for chunked getLogs
+- [#2446] Add parameter for subkey generation to overwrite origin URL
 
 ### Changed
 - [#2409] Lower default payment expiration to 1.1 × reveal timeout
@@ -13,6 +14,7 @@
 [#2409]: https://github.com/raiden-network/light-client/issues/2409
 [#2417]: https://github.com/raiden-network/light-client/pull/2417
 [#2444]: https://github.com/raiden-network/light-client/issues/2444
+[#2446]: https://github.com/raiden-network/light-client/issues/2446
 
 ## [0.14.0] - 2020-11-25
 ### Fixed